* 'master' of https://github.com/processwire/processwire: (273 commits)
  Bump version to 3.0.123
  Update for processwire/processwire-issues#408 InputfieldSelector count() to wireCount()
  Add $options argument support to WireMailTools::new(), i.e. $mail->new(array $options)
  Add additional mbstring check to Sanitizer and add two new whitespace reduction methods (for internal use)
  Fix issue processwire/processwire-issues#724 UTF-8 pagenames in non-default language
  Update for processwire/processwire-issues#767
  Bump version to 3.0.122 plus some other minor adjustments
  Update WireArray::import() to support duplicate items when duplicate checking is turned off per processwire/processwire-issues#767
  Fix issue processwire/processwire-issues#766 auto-remove UTF-8 value of `&#8232` line-seperator entity in $sanitizer->text() function
  Add @horst-n fix for processwire/processwire-issues#715
  Fix issue processwire/processwire-issues#763 for ProcessRole, plus this commit also contains some minor tweaks to ProcessPageLister
  Fix issue with Safari column widths in AdminThemeUikit that use showIf conditions per processwire/processwire-issues#480
  Fix issue processwire/processwire-issues#756 make link modal files selection not require being open before it can be populated with files from page selection
  Fix issue processwire/processwire-issues#723
  Fix issue processwire/processwire-issues#749
  A couple of updates to account for new PHP 7.2/7.3 notices per processwire/processwire-issues#408
  Various minor tweaks/updates and bump version to 3.0.121
  Update locale warning message in ProcessLogin per processwire/processwire-issues#732
  Update phpdoc @return statement for Page::index()
  Additional updates for processwire/processwire-issues#751 plus some enhancements to PageFinder
  ...

+++ wire/core/WireTempDir.php
@@ -3,7 +3,7 @@
 /**
  * ProcessWire Temporary Directory Manager
  *
- * ProcessWire 3.x, Copyright 2017 by Ryan Cramer
+ * ProcessWire 3.x, Copyright 2018 by Ryan Cramer
  * https://processwire.com
  *
  */
@@ -85,13 +85,9 @@ class WireTempDir extends Wire {
 	 * 
 	 */
 	public function createName($prefix = '') {
-		$pass = new Password();
-		$this->wire($pass);
-		$len = mt_rand(10, 30);
-		$name = microtime() . '.' . $pass->randomBase64String($len, true);
-		$a = explode('.', $name);
-		shuffle($a);
-		$name = $prefix . implode('O', $a);
+		$random = new WireRandom();
+		$len = $random->integer(10, 20);
+		$name = $prefix . str_replace(' ', 'T', microtime()) . 'R' . $random->alphanumeric($len);
 		$this->createdName = $name;
 		return $name;
 	}
@@ -206,7 +202,7 @@ class WireTempDir extends Wire {
 			return true;
 		}
 	
-		if(is_dir($this->tempDir)) {
+		if($this->tempDir && is_dir($this->tempDir)) {
 			// remove temporary directory created by this instance
 			if(!$this->rmdir($this->tempDir, true)) {
 				$this->log("$errorMessage: $this->tempDir");
@@ -214,7 +210,7 @@ class WireTempDir extends Wire {
 			}
 		}
 	
-		if(is_dir($this->tempDirRoot)) {
+		if($this->tempDirRoot && is_dir($this->tempDirRoot)) {
 			if($this->createdName && strpos($this->tempDirRoot, "/.$this->createdName")) {
 				// if tempDirRoot is just for this PW instance, we can remove it now
 				$this->rmdir($this->tempDirRoot, true);
@@ -224,7 +220,7 @@ class WireTempDir extends Wire {
 			}
 		}
 		
-		if(!$classRuns && is_dir($this->classRoot)) {
+		if(!$classRuns && $this->classRoot && is_dir($this->classRoot)) {
 			$this->removeExpiredDirs($this->classRoot, 86400); 
 		}
 	
@@ -249,7 +245,6 @@ class WireTempDir extends Wire {
 		if(!is_int($maxAge)) $maxAge = $this->tempDirMaxAge;
 		
 		$numSubdirs = 0;
-		$pathname = '';
 		$oldestAllowedFileTime = time() - $maxAge;
 		$success = true;
 		
@@ -264,10 +259,8 @@ class WireTempDir extends Wire {
 			}
 			
 			// old dir found: check times on files/dirs within that dir
-			$pathname = rtrim($dir->getPathname(), "/\\") . DIRECTORY_SEPARATOR;
-			
-			// if our .wtd identifier file isn't present, ignore the directory
-			if(!is_file($pathname . self::hiddenFileName)) continue;
+			$pathname = $this->wire('files')->unixDirName($dir->getPathname());
+			if(!$this->isTempDir($pathname)) continue;
 			
 			$removeDir = true;
 			$newestFileTime = $this->getNewestModTime($pathname); 
@@ -275,7 +268,7 @@ class WireTempDir extends Wire {
 		
 			if($removeDir) {
 				if(!$this->rmdir($pathname, true)) {
-					$this->log("Unable to remove: $path");
+					$this->log("Unable to remove (B): $pathname");
 					$success = false;
 				}
 			} else {
@@ -283,12 +276,12 @@ class WireTempDir extends Wire {
 			}
 		}
 
-		if(!$numSubdirs) {
+		if(!$numSubdirs && $path != $this->classRoot && $this->isTempDir($path)) {
 			// if no subdirectories, we can remove the root
 			if($this->rmdir($path, true)) {
 				$success = true;
 			} else {
-				$this->log("Unable to remove: $pathname");
+				$this->log("Unable to remove (A): $path");
 				$success = false;
 			}
 		}
@@ -353,9 +346,12 @@ class WireTempDir extends Wire {
 	 * 
 	 */
 	protected function mkdir($dir, $recursive = false) {
-		if($this->wire('files')->mkdir($dir, $recursive)) {
-			$dir = rtrim($dir, "/\\") . DIRECTORY_SEPARATOR;
-			file_put_contents($dir . self::hiddenFileName, time());
+		/** @var WireFileTools $files */
+		$files = $this->wire('files');
+		$dir = $files->unixDirName($dir);
+		if($files->mkdir($dir, $recursive)) {
+			$files->filePutContents($dir . self::hiddenFileName, time());
+			return true;
 		} else {
 			return false;
 		}
@@ -370,9 +366,31 @@ class WireTempDir extends Wire {
 	 * 
 	 */
 	protected function rmdir($dir, $recursive = false) {
-		$dir = rtrim($dir, "/\\") . DIRECTORY_SEPARATOR;
-		if(!is_file($dir . self::hiddenFileName)) return false;
-		unlink($dir . self::hiddenFileName);
-		return $this->wire('files')->rmdir($dir, $recursive);
+		$dir = $this->wire('files')->unixDirName($dir);
+		if(!strlen($dir) || !is_dir($dir)) return true;
+		if(!$this->isTempDir($dir)) return false;
+		if(is_file($dir . self::hiddenFileName)) $this->wire('files')->unlink($dir . self::hiddenFileName, true);
+		return $this->wire('files')->rmdir($dir, $recursive, true);
 	}
+
+	/**
+	 * Is given directory/path created by this class?
+	 * 
+	 * @param string $dir
+	 * @return bool
+	 * 
+	 */
+	protected function isTempDir($dir) {
+		$files = $this->wire('files');
+		if(!strlen($dir) || !is_dir($dir)) {
+			// if given a non-directory return false
+			return false;
+		}
+		if($this->classRoot && $files->fileInPath($dir, $this->classRoot)) {
+			// dir is within classRoot path
+			return true;
+		}
+		return false;
+	}
+
 }
